+++ packages/services/api/src/modules/integrations/providers/github-integration-manager.ts
@@ -4,7 +4,9 @@ import { Octokit } from '@octokit/core';
 import { RequestError } from '@octokit/request-error';
 import type { GitHubIntegration } from '../../../__generated__/types';
 import { HiveError } from '../../../shared/errors';
-import { Session } from '../../auth/lib/authz';
+import { AuthManager } from '../../auth/providers/auth-manager';
+import { OrganizationAccessScope } from '../../auth/providers/organization-access';
+import { ProjectAccessScope } from '../../auth/providers/scopes';
 import { Logger } from '../../shared/providers/logger';
 import { OrganizationSelector, ProjectSelector, Storage } from '../../shared/providers/storage';
 
@@ -27,7 +29,7 @@ export class GitHubIntegrationManager {
 
   constructor(
     logger: Logger,
-    private session: Session,
+    private authManager: AuthManager,
     private storage: Storage,
     @Inject(GITHUB_APP_CONFIG) private config: GitHubApplicationConfig | null,
   ) {
@@ -63,12 +65,9 @@ export class GitHubIntegrationManager {
       input.organizationId,
       input.installationId,
     );
-    await this.session.assertPerformAction({
-      action: 'gitHubIntegration:modify',
-      organizationId: input.organizationId,
-      params: {
-        organizationId: input.organizationId,
-      },
+    await this.authManager.ensureOrganizationAccess({
+      ...input,
+      scope: OrganizationAccessScope.INTEGRATIONS,
     });
     this.logger.debug('Updating organization');
     await this.storage.addGitHubIntegration({
@@ -79,13 +78,9 @@ export class GitHubIntegrationManager {
 
   async unregister(input: OrganizationSelector): Promise<void> {
     this.logger.debug('Removing GitHub integration (organization=%s)', input.organizationId);
-
-    await this.session.assertPerformAction({
-      action: 'gitHubIntegration:modify',
-      organizationId: input.organizationId,
-      params: {
-        organizationId: input.organizationId,
-      },
+    await this.authManager.ensureOrganizationAccess({
+      ...input,
+      scope: OrganizationAccessScope.INTEGRATIONS,
     });
     this.logger.debug('Updating organization');
     await this.storage.deleteGitHubIntegration({
@@ -198,12 +193,9 @@ export class GitHubIntegrationManager {
       return null;
     }
 
-    await this.session.assertPerformAction({
-      action: 'gitHubIntegration:modify',
+    await this.authManager.ensureOrganizationAccess({
       organizationId: organization.id,
-      params: {
-        organizationId: organization.id,
-      },
+      scope: OrganizationAccessScope.INTEGRATIONS,
     });
 
     return organization;
@@ -427,12 +419,9 @@ export class GitHubIntegrationManager {
   }
 
   async enableProjectNameInGithubCheck(input: ProjectSelector) {
-    await this.session.assertPerformAction({
-      action: 'gitHubIntegration:modify',
-      organizationId: input.organizationId,
-      params: {
-        organizationId: input.organizationId,
-      },
+    await this.authManager.ensureProjectAccess({
+      ...input,
+      scope: ProjectAccessScope.SETTINGS,
     });
 
     const project = await this.storage.getProject(input);
