+++ packages/services/api/src/modules/integrations/providers/slack-integration-manager.ts
@@ -1,6 +1,9 @@
 import { Injectable, Scope } from 'graphql-modules';
 import { AccessError } from '../../../shared/errors';
-import { Session } from '../../auth/lib/authz';
+import { AuthManager } from '../../auth/providers/auth-manager';
+import { OrganizationAccessScope } from '../../auth/providers/organization-access';
+import { ProjectAccessScope } from '../../auth/providers/project-access';
+import { TargetAccessScope } from '../../auth/providers/target-access';
 import { CryptoProvider } from '../../shared/providers/crypto';
 import { Logger } from '../../shared/providers/logger';
 import {
@@ -20,7 +23,7 @@ export class SlackIntegrationManager {
 
   constructor(
     logger: Logger,
-    private session: Session,
+    private authManager: AuthManager,
     private storage: Storage,
     private crypto: CryptoProvider,
   ) {
@@ -35,12 +38,9 @@ export class SlackIntegrationManager {
     },
   ): Promise<void> {
     this.logger.debug('Registering Slack integration (organization=%s)', input.organizationId);
-    await this.session.assertPerformAction({
-      action: 'slackIntegration:modify',
-      organizationId: input.organizationId,
-      params: {
-        organizationId: input.organizationId,
-      },
+    await this.authManager.ensureOrganizationAccess({
+      ...input,
+      scope: OrganizationAccessScope.INTEGRATIONS,
     });
     this.logger.debug('Updating organization');
     await this.storage.addSlackIntegration({
@@ -51,12 +51,9 @@ export class SlackIntegrationManager {
 
   async unregister(input: OrganizationSelector): Promise<void> {
     this.logger.debug('Removing Slack integration (organization=%s)', input.organizationId);
-    await this.session.assertPerformAction({
-      action: 'slackIntegration:modify',
-      organizationId: input.organizationId,
-      params: {
-        organizationId: input.organizationId,
-      },
+    await this.authManager.ensureOrganizationAccess({
+      ...input,
+      scope: OrganizationAccessScope.INTEGRATIONS,
     });
     this.logger.debug('Updating organization');
     await this.storage.deleteSlackIntegration({
@@ -108,12 +105,9 @@ export class SlackIntegrationManager {
           selector.organizationId,
           selector.context,
         );
-        await this.session.assertPerformAction({
-          action: 'slackIntegration:modify',
-          organizationId: selector.organizationId,
-          params: {
-            organizationId: selector.organizationId,
-          },
+        await this.authManager.ensureOrganizationAccess({
+          ...selector,
+          scope: OrganizationAccessScope.INTEGRATIONS,
         });
         break;
       }
@@ -124,13 +118,9 @@ export class SlackIntegrationManager {
           selector.projectId,
           selector.context,
         );
-        await this.session.assertPerformAction({
-          action: 'alert:modify',
-          organizationId: selector.organizationId,
-          params: {
-            organizationId: selector.organizationId,
-            projectId: selector.projectId,
-          },
+        await this.authManager.ensureProjectAccess({
+          ...selector,
+          scope: ProjectAccessScope.ALERTS,
         });
         break;
       }
@@ -142,15 +132,9 @@ export class SlackIntegrationManager {
           selector.targetId,
           selector.context,
         );
-        await this.session.assertPerformAction({
-          action: 'schemaVersion:publish',
-          organizationId: selector.organizationId,
-          params: {
-            organizationId: selector.organizationId,
-            projectId: selector.projectId,
-            targetId: selector.targetId,
-            serviceName: null,
-          },
+        await this.authManager.ensureTargetAccess({
+          ...selector,
+          scope: TargetAccessScope.REGISTRY_WRITE,
         });
         break;
       }
